requirements.txt
asgiref==3.5.2
astroid==2.11.5
bcrypt==3.2.2
black==22.3.0
cffi==1.15.0
click==8.1.3
dill==0.3.5.1
Django==4.0.5
django-environ==0.8.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.14.1
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==4.2.0
wrapt==1.14.1
